Organising the reset menu
